+++ CHANGELOG.md
@@ -31,7 +31,7 @@ You can check your current version with the following command:
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
 ## 0.37.1a5
 
-**March 20, 2024**
+**March 21, 2024**
 
 - Run test pipeline on Python versions 3.9 to 3.12
 - Removed upper bound for Python 3.12